# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Hôtel Artichaut",
        "version": "1.0"
    },
    "paths": {
...
        "api/user": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher tous les utilisateurs",
                "operationId": "2c54654f7409a9b71065e0a53fd95671",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "5d7c0fda78d3186150f7d4fa7182f7f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "email",
                                    "password",
                                    "phone"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "emailBis": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "phoneBis": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "maximum": 2,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "api/user/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher un utilisateur par ID",
                "operationId": "5ca97d73e49b988e6f6c8f9d921082b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'utilisateur"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "920a079a12b7def8208422e5c4e518bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "email",
                                    "password",
                                    "phone"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "emailBis": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "phoneBis": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "maximum": 2,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "c23411da8fb9c574e7ef85bf1638fec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        }
    },
...
}